Move docker file, compute docker image name dynamically
